Scenes for everyone!
3 scene folders - Final, GameJam Testing, and Main Scene Testing. Final
folder is the scenes that will be in the real game (the
MainScene(Tristan) is the one that Tristan was using during the gamejam,
and that we ended up using for the final game. GameJam testing is all of
our non-final scenes during the gamejam, which I have not modified. The
main scene testing folder is for me, Roger, and Mike. They are all
duplicates of the MainScene(Tristan) to use for testing.